feat: 1-45 숫자인 targetNumbers 일급 컬렉션 적용
