+++ src/main/java/Application.java
@@ -1,6 +1,7 @@
 import domain.lotteryService.LotteryMachine;
 import domain.lotteryStore.Lotteries;
 import domain.lotteryStore.LotteryStore;
+import domain.lotteryStore.TargetNumbers;
 import ui.Printer;
 import ui.receiver.InputValidator;
 import ui.receiver.Receiver;
@@ -14,7 +15,7 @@ public class Application {
         InputValidator.validatePurchasePrice(purchasePrice);
         int purchasedCount = printer.printPurchasedCount(purchasePrice);
 
-        LotteryStore lotteryStore = new LotteryStore();
+        LotteryStore lotteryStore = new LotteryStore(new TargetNumbers());
         Lotteries lotteries = lotteryStore.createLotteries(purchasedCount);
         printer.printPurchasedLotteries(lotteries);
